// src/constants/types/dataPostsType.ts
export interface SearchPostsTypeInterface {
  id: string;
  nameEn: string;
  nameEs: string;
  nameDe: string;
  nameIt: string;
  nameFr: string;
  nameRu: string;
  path: string;
  pathToShop: string;
  type: string;
  secondType: string;
  keyWords: string;
  dateCreate: number;
  totalReads: number;
  like: number;
  productName: string;
  image: string;
  descriptionsDe: string;
  descriptionsEn: string;
  descriptionsEs: string;
  descriptionsIt: string;
  descriptionsFr: string;
  descriptionsRu: string;
}

export interface DataPostsTypeInterface {
  id: string;
  nameEn: string;
  nameEs: string;
  nameDe: string;
  nameIt: string;
  nameFr: string;
  nameRu: string;
  path: string;
  pathToShop: string;
  type: string;
  secondType: string;
  keyWords: string;
  dateCreate: number;
  totalReads: number;
  like: number;
  productName: string;
  image: string;
  descriptionsDe: string;
  descriptionsEn: string;
  descriptionsEs: string;
  descriptionsIt: string;
  descriptionsFr: string;
  descriptionsRu: string;
}